pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sbatchman"
version = "0.1.3"
authors = [
  { name="Hicrest laboratory", email="hicrest@unitn.it" },
]
description = "A utility to create, launch and monitor code experiments on SLURM, PBS, or local machines."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer[all]",
    "textual",
    "rich",
    "PyYAML",
    "platformdirs",
]

[project.urls]
"Homepage" = "https://github.com/LorenzoPichetti/SbatchMan"
"Bug Tracker" = "https://github.com/LorenzoPichetti/SbatchMan/issues"

[project.scripts]
sbatchman = "sbatchman.cli:app"